sloppy merge of two constants files
