+++ src/main/java/frc/robot/commands/auto/FiveBallAuto.java
@@ -33,7 +33,7 @@ import com.pathplanner.lib.commands.PPSwerveControllerCommand;
 
 import edu.wpi.first.math.util.Units;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
-import frc.robot.Constants.DriveConstants;
+import frc.robot.lists.Constants.DriveConstants;
 import frc.robot.subsystems.Drivetrain;
 
 public class FiveBallAuto extends SequentialCommandGroup {
