+++ src/main/java/frc/robot/commands/swerve/DriveModuleTest.java
@@ -3,7 +3,7 @@ package frc.robot.commands.swerve;
 // Imports
 import edu.wpi.first.math.MathUtil;
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.Constants.DriveConstants.ModulePosition;
+import frc.robot.lists.Constants.DriveConstants.ModulePosition;
 import frc.robot.subsystems.Drivetrain;
 import java.util.function.DoubleSupplier;
